cambio nombre arcchivo funciones stock
